use bigdecimal in constructor, normalize spaces

+++ backend/numbers-to-words-backend/src/main/java/io/github/furnapso/numberstowordsbackend/Test.java
@@ -2,8 +2,10 @@ package io.github.furnapso.numberstowordsbackend;
 
 import io.github.furnapso.numberstowordsbackend.parse.NumberToWordParser;
 
+import java.math.BigDecimal;
+
 public class Test {
     public static void main(String[] args) {
-        System.out.println(new NumberToWordParser(384934.52f).parse());
+        System.out.println(new NumberToWordParser(BigDecimal.valueOf(384934.52)).parse());
     }
 }
